modify data type

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -4,10 +4,12 @@
  * elements of a list_t list
  *
  * @h: variable of type pointer to struct list_t
+ *
+ * Return: number of nodos of the linked lists
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
